requirements/base.txt
anyio==3.6.2
click==8.1.3
dnspython==2.2.1
fastapi==0.85.2
graphql-core==3.2.3
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.2
pymongo==4.3.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
strawberry-graphql==0.138.1
typing_extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.4
bcrypt==4.0.1